requirements.txt
python-telegram-bot==13.15
msrest==0.7.1
azure-devops==6.0.0
i18n==0.10.7
